#1216 Feedback/making sure everything still works
