+++ src/frontend/src/components/ActionsMenu.tsx
@@ -2,10 +2,17 @@ import { Box } from '@mui/system';
 import { ReactElement, useState } from 'react';
 import { NERButton } from './NERButton';
 import { ArrowDropDown } from '@mui/icons-material';
-import { Menu } from '@mui/material';
+import { ListItemIcon, Menu, MenuItem } from '@mui/material';
+
+export type ButtonInfo = {
+  title: string;
+  onClick: () => void;
+  disabled?: boolean;
+  icon?: ReactElement;
+};
 
 interface ActionsMenuProps {
-  buttons: ReactElement<any, any>[];
+  buttons: ButtonInfo[];
 }
 
 const ActionsMenu: React.FC<ActionsMenuProps> = ({ buttons }) => {
@@ -33,7 +40,10 @@ const ActionsMenu: React.FC<ActionsMenuProps> = ({ buttons }) => {
       </NERButton>
       <Menu open={dropdownOpen} anchorEl={anchorEl} onClose={handleDropdownClose}>
         {buttons.map((button, index) => (
-          <Box key={index}>{button}</Box>
+          <MenuItem key={index} onClick={button.onClick} disabled={button.disabled}>
+            <ListItemIcon>{button.icon}</ListItemIcon>
+            {button.title}
+          </MenuItem>
         ))}
       </Menu>
     </Box>
