+++ src/frontend/src/utils/routes.ts
@@ -13,7 +13,7 @@ const CREDITS = `/credits`;
 
 /**************** Finance Section ****************/
 const FINANCE = `/finance`;
-const REIMBURSEMENT_REQUESTS = '/reimbursement-requests';
+const REIMBURSEMENT_REQUESTS = FINANCE + '/reimbursement-requests';
 const REIMBURSEMENT_REQUEST_BY_ID = REIMBURSEMENT_REQUESTS + `/:id`;
 const REIMBURSEMENT_REQUEST_EDIT = REIMBURSEMENT_REQUEST_BY_ID + `/edit`;
 const NEW_REIMBURSEMENT_REQUEST = REIMBURSEMENT_REQUESTS + `/new`;
